+++ FrontEnd/src/index.js
@@ -5,10 +5,12 @@ import App from "./App";
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import Settings from './pages/Settings/Settings';
-import { SettingsProvider } from "./pages/Settings/SettingsContext";
 import Pricing from './pages/Pricing/Pricing';
 import Login from './pages/Login/Login';
 import { apiFetch } from "./utils/authUtils";
+
+import { SettingsProvider } from "./pages/Settings/SettingsContext";
+import { SelectionProvider } from "./pages/Messages/SelectionContext";
 
 import 'highlight.js/styles/atom-one-dark.css';
 import Messages from "./pages/Messages/Messages";
@@ -42,20 +44,22 @@ function RootApp() {
 
   return (
     <SettingsProvider>
-      <BrowserRouter>
-        <Routes>
-          {/* Public Routes */}
-          <Route path="/login" element={<Login />} />
+      <SelectionProvider>
+        <BrowserRouter>
+          <Routes>
+            {/* Public Routes */}
+            <Route path="/login" element={<Login />} />
 
-          {/* Protected Routes */}
-          <Route path="/" element={isAuthenticated ? <App /> : <Navigate to="/login" />} />
-          <Route path="/settings" element={isAuthenticated ? <Settings /> : <Navigate to="/login" />} />
-          <Route path="/pricing" element={isAuthenticated ? <Pricing /> : <Navigate to="/login" />} />
-          <Route path="/messages" element={isAuthenticated ? <Messages /> : <Navigate to="/login" />} />
-        </Routes>
-      </BrowserRouter>
+            {/* Protected Routes */}
+            <Route path="/messages" element={isAuthenticated ? <Messages /> : <Navigate to="/login" />} />
+            <Route path="/" element={isAuthenticated ? <App /> : <Navigate to="/login" />} />
+            <Route path="/settings" element={isAuthenticated ? <Settings /> : <Navigate to="/login" />} />
+            <Route path="/pricing" element={isAuthenticated ? <Pricing /> : <Navigate to="/login" />} />
+            
+          </Routes>
+        </BrowserRouter>
+      </SelectionProvider>
     </SettingsProvider>
-    
   );
 }